requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0

